+++ rye/src/cli/mod.rs
@@ -72,12 +72,20 @@ enum Command {
     Version(version::Args),
     #[command(hide = true)]
     List(list::Args),
+    #[command(hide = true)]
+    Shell(shell::Args),
 }
 
 pub mod list {
     /// There is no real list command yet.
     ///
     /// Use rye show --installed-deps instead
+    #[derive(clap::Parser, Debug)]
+    pub struct Args {}
+}
+
+pub mod shell {
+    /// The shell command was removed.
     #[derive(clap::Parser, Debug)]
     pub struct Args {}
 }
@@ -132,6 +140,16 @@ pub fn execute() -> Result<(), Error> {
             // until we have a proper list command, make it error with what the
             // user should be using instead.
             bail!("unknown command. Maybe you mean rye show --installed-deps");
+        }
+        Command::Shell(..) => {
+            bail!(
+                "unknown command. The shell command was removed. Activate the virtualenv instead with '{}' instead.",
+                if cfg!(windows) {
+                    ".venv\\Scripts\\activate"
+                } else {
+                    ". .venv/bin/activate"
+                }
+            );
         }
     }
 }